// dsl/packages/cli/test/test-specs/gen-out/client/typescript-client/src/model/ScalarRecord.ts
// Generated by RSD - Do not modify
import { isZoneId, type ZoneId } from './Scalars.js';
import { checkOptProp, checkProp, createTypedArrayGuard, isNull, isRecord, isString, propListValue, propValue } from '../_type-utils.js';

export type ScalarRecord = {
	readonly value: ZoneId;
	readonly value_Null: ZoneId | null;
	readonly value_Opt?: ZoneId;
	readonly value_Opt_Null?: ZoneId | null;
	readonly list: ZoneId[];
	readonly list_Null: ZoneId[] | null;
	readonly list_Opt?: ZoneId[];
	readonly list_Opt_Null?: ZoneId[] | null;
};

export function isScalarRecord(value: unknown): value is ScalarRecord {
	return isRecord(value) &&
		checkProp(value, 'value', isZoneId) &&
		(checkProp(value, 'value_Null', isNull) || checkProp(value, 'value_Null', isZoneId)) &&
		checkOptProp(value, 'value_Opt', isZoneId) &&
		(checkOptProp(value, 'value_Opt_Null', isNull) || checkOptProp(value, 'value_Opt_Null', isZoneId)) &&
		checkProp(value, 'list', createTypedArrayGuard(isZoneId)) &&
		(checkProp(value, 'list_Null', isNull) || checkProp(value, 'list_Null', createTypedArrayGuard(isZoneId))) &&
		checkOptProp(value, 'list_Opt', createTypedArrayGuard(isZoneId)) &&
		(checkOptProp(value, 'list_Opt_Null', isNull) || checkOptProp(value, 'list_Opt_Null', createTypedArrayGuard(isZoneId)));
}

export function ScalarRecordFromJSON($value: Record<string, unknown>): ScalarRecord {
	const value = propValue('value', $value, isString);
	const value_Null = propValue('value_Null', $value, isString, 'null');
	const value_Opt = propValue('value_Opt', $value, isString, 'optional');
	const value_Opt_Null = propValue('value_Opt_Null', $value, isString, 'optional_null');
	const list = propListValue('list', $value, isString);
	const list_Null = propListValue('list_Null', $value, isString, 'null');
	const list_Opt = propListValue('list_Opt', $value, isString, 'optional');
	const list_Opt_Null = propListValue('list_Opt_Null', $value, isString, 'optional_null');
	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}

export function ScalarRecordToJSON($value: ScalarRecord): Record<string, unknown> {
	const value = $value.value;
	const value_Null = $value.value_Null;
	const value_Opt = $value.value_Opt;
	const value_Opt_Null = $value.value_Opt_Null;
	const list = $value.list;
	const list_Null = $value.list_Null;
	const list_Opt = $value.list_Opt;
	const list_Opt_Null = $value.list_Opt_Null;

	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}